Stereoscopic Projection

+++ libs/gm/src/gm/color/color_convertor.cpp
@@ -4,10 +4,11 @@
 
 #include <gm/color/color_convertor.h>
 
+
 unsigned int colorGLToSTD(float x) {
     return COLOR_STD_MAX * x;
 }
 
 float colorSTDToGL(unsigned int x) {
-    return (x / COLOR_STD_MAX * COLOR_GL_MAX);
+    return (x / (float)COLOR_STD_MAX * COLOR_GL_MAX);
 }
